Explicitly set `localhost` when loading dumps
The problem occurs when the default are modified. Sockets can't be
created in `/var/run` as the permissions are wrong, so we can patch the
`unix_socket_directory` to something else. This works for most commands
as they will reference the data directory, but `psql` does not. It
will only look for the socket in `/var/run`, so we explicitly set the
hostname to avoid the unix socket lookup.